requirements.txt
customtkinter
cryptography
pynput
pyinstaller
requests
tkinterdnd2